refactor: move meme-generator files to feature folder
